Refactor auth to use Astro Actions and form submit
Replace raw API routes with Astro Actions for type-safe server
communication. Convert login page from button click handler to
form submit. Update logout to use action as well.
